python/requirements-mqtt_sparkplug.txt
paho-mqtt==1.4.0
google==2.0.2
protobuf==3.7.0
